lost data

+++ modules/task_3/poletueva_a_linear_filtering/poletueva_a_linear_filtering.cpp
@@ -11,7 +11,6 @@
 std::vector<int> LinFilter(const std::vector <int>& a, int rows, int cols) {
   if (rows < 0 || cols < 0)
     throw Error;
-
   std::vector<int> res = a;
   std::vector<double> h = { 0.0625, 0.125, 0.0625, 0.125, 0.25, 0.125, 0.0625, 0.125, 0.0625 };
 
@@ -20,7 +19,7 @@ std::vector<int> LinFilter(const std::vector <int>& a, int rows, int cols) {
     for (int x = 1; x < rows - 1; x++)
       for (int i = -r; i <= r; i++)
         for (int j = -r; j <= r; j++)
-          res[x + y *rows] += a[x + i + (y + j)*rows] * h[i + 1 + (j + 1) * 3];
+          res[x + y *rows] += static_cast<int>( a[x + i + (y + j)*rows] * h[i + 1 + (j + 1) * 3]);
   return res;
 }
